refactor: finalize agent ui structure
