+++ packages/browseros-agent/apps/agent/components/elements/available-tabs.hooks.ts
