+++ packages/browseros-agent/apps/agent/screens/newtab/index/lib/search-suggestions/search-suggestions.hooks.ts
@@ -1,6 +1,6 @@
 import { useEffect, useState } from 'react'
 import useSWR from 'swr'
-import { getSearchSuggestions } from './getSearchSuggestions'
+import { getSearchSuggestions } from './search-suggestions'
 
 interface UseSearchSuggestionsArgs {
   query: string
