+++ packages/browseros-agent/apps/agent/screens/newtab/index/lib/search-suggestions/search-suggestions.ts
